@@ -1,10 +1,7 @@
 <script lang="ts">
-  let count: number = 0
-  const increment = () => {
-    count += 1
-  }
+  let count: number = $state(0);
 </script>
 
-<button on:click={increment}>
+<button onclick={() => count++}>
   count is {count}
 </button>
