@@ -1,8 +1,6 @@
 import './app.css'
 import App from './App.svelte'
-
-const app = new App({
-  target: document.getElementById('app')!,
-})
+import { mount } from 'svelte'
+const app = mount(App, { target: document.getElementById('app')! })
 
 export default app
